List Addition With .items() (Python2/Python3)
* Typecast the returned dict_items from Python3 into list

Local testing with sample dictionaries in Python 2 and Python 3 via Docker shows it resolves the issue. It does cause an aditional list to be created in Python 2 because it returns a list already. The overhead for the limited set of users still on Python 2 this is likely not a concern.